% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcop.R
\name{hcop}
\alias{hcop}
\title{Retrieves the hcop ortholog map data.frame}
\usage{
hcop(cache = TRUE)
}
\arguments{
\item{cache}{If \code{TRUE} (default), will store the data.frame internally so
that it doesn't have to be re-read the next time the function is called}
}
\value{
a data.frame with human to species ortholog mapping info
}
\description{
This is the data.frame that is used to map the collection data from its
source/raw input (human, entrez), to ensembl or other species.
}
\details{
HCOPS is the HGNC Comparison of Orthology Predictions.
}
\seealso{
https://www.genenames.org/help/hcop/
}
