% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb_retrieve.R
\name{msigdb_load}
\alias{msigdb_load}
\title{Loads the reference msigdb collection data.frame}
\usage{
msigdb_load(version = NULL, cache = TRUE)
}
\arguments{
\item{version}{Specifies the version of the database to load. If \code{NULL}
(the default), the latest version is retrieved. To see what other versions
are available, refer to the \code{version} column in the \code{\link[=msigdb_versions]{msigdb_versions()}}
data.frame}
}
\value{
The full data.frame of gene set collections for the give version.
The version is added as a \code{msigdb_version} attribute to the returned
result.
}
\description{
These are all-human identifiers.
}
\details{
Info for v7: http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/MSigDB_v7.0_Release_Notes
}
