% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb_versions.R
\name{msigdb_versions}
\alias{msigdb_versions}
\title{Returns a table of available msigdb versions within this package.}
\usage{
msigdb_versions(
  dat.dir = NULL,
  dir.pattern = "msigdb-collection_v.*rds$",
  version.regex = "(\\\\d+)\\\\.(\\\\d+)\\\\.(\\\\d+)?"
)
}
\arguments{
\item{dat.dir}{the directory that holds the version data. By default it
looks at the package's internal \code{"extdata"} pacakge for the collection
data.}

\item{dir.pattern}{what the misgdb files look like}

\item{version.regex}{how the versions are encoded}
}
\value{
a tibble of msigdb version info
}
\description{
The parameters in this function really shouldn't be used. By default
it checks the internal package directory to see what msigdb definitions
are included.
}
