% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goslim.R
\name{slim_ids}
\alias{slim_ids}
\title{Returns the ID's that map to a specific GO subset / slim}
\usage{
slim_ids(name = "generic")
}
\value{
a vector of GO ids
}
\description{
Returns the ID's that map to a specific GO subset / slim
}
