% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_info.R
\name{species_info}
\alias{species_info}
\alias{species_lookup}
\title{Species information table and lookup}
\usage{
species_info(hcop_only = FALSE)

species_lookup(query, stable = species_info(FALSE), ignore.case = TRUE)
}
\arguments{
\item{hcop_only}{Only retrun species information for organisms that are
in the \code{\link[=hcop]{hcop()}} database.}
}
\value{
A single-row tibble with (numeric) species id, name, etc.
}
\description{
\code{species_info()} returns a table of ncbi species id's, names, common name,
etc. \code{species_lookup()} performs fuzzy matches against the table to return
the species-level info for a query. The fuzzy match must start left-to-right.
}
\examples{
species_lookup("human")
species_lookup("scer")
species_lookup("hsapiens")
}
