% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb_retrieve.R
\name{msigdb_retrieve}
\alias{msigdb_retrieve}
\title{Retrieves a msgidb collection data.frame for the given species}
\usage{
msigdb_retrieve(species, collections = NULL, id_type = c("ensembl",
  "entrez", "symbol"))
}
\arguments{
\item{species}{the name of the species (human, mouse, etc.) see entries
in GeneSetDb.MSigDB:::.species_tbl()[\link{"common_name}] (+ "human")}

\item{collections}{character of MSigDb collections ("h", "c1", ..., "c7")}

\item{id_type}{"ensembl", "entrez", or "symbol"}
}
\description{
Retrieves a msgidb collection data.frame for the given species
}
